examples/sandbox/pyproject.toml
[project]
name = "easyagent-sandbox-example"
version = "0.1.0"
description = "Sandbox example environment for easyagent with opensandbox-server"
requires-python = ">=3.13"
dependencies = [
    "easyagent",
    "opensandbox-server",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
easyagent = { path = "../..", editable = true }
